backend/requirements.txt
Flask==2.1.1
psycopg2==2.9.3
flask-cors==3.0.10
python-dotenv==0.19.2
Werkzeug==2.0.3
SQLAlchemy==1.4.31
